requirements.txt
tensorflow
valohai-utils
valohai-yaml
pandas
numpy 
sklearn